fix: rubocop refactor
