+++ Rakefile
@@ -1,4 +1,6 @@
-require "bundler/gem_tasks"
+# frozen_string_literal: true
+
+require 'bundler/gem_tasks'
 load 'lib/tasks/bsb_tasks.rake'
 
 require 'rake/testtask'
@@ -7,5 +9,5 @@ Rake::TestTask.new do |t|
   t.libs << 'test'
 end
 
-desc "Run tests"
-task :default => :test
+desc 'Run tests'
+task default: :test
