+++ bsb.gemspec
@@ -20,7 +20,7 @@ Gem::Specification.new do |spec|
   spec.require_paths = ['lib']
 
   spec.add_development_dependency 'bundler', '~> 2.0'
-  spec.add_development_dependency 'rake', '~> 13.0'
   spec.add_development_dependency 'net-ftp', '~> 0.1.3'
+  spec.add_development_dependency 'rake', '~> 13.0'
   spec.add_dependency 'activemodel'
 end
