+++ lib/auspaynet/client.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require 'net/ftp'
 
 module Auspaynet
