+++ lib/bsb.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require 'bsb/version'
 require 'json'
 require 'bsb_number_validator'
@@ -8,7 +10,8 @@ module BSB
       bsb = normalize(number)
       array = data_hash[bsb]
       return nil if array.nil?
-      result = {
+
+      {
         bsb: bsb,
         mnemonic: array[0],
         bank_name: bank_name(bsb),
@@ -18,9 +21,9 @@ module BSB
         state: array[4],
         postcode: array[5],
         flags: {
-          paper: (array[6][0] == "P"),
-          electronic: (array[6][1] == "E"),
-          high_value: (array[6][2] == "H")
+          paper: (array[6][0] == 'P'),
+          electronic: (array[6][1] == 'E'),
+          high_value: (array[6][2] == 'H')
         }
       }
     end
@@ -29,7 +32,7 @@ module BSB
       bank_list.each do |prefix, bank_name|
         return bank_name if bsb.start_with? prefix
       end
-      return nil
+      nil
     end
 
     def normalize(str)
@@ -37,12 +40,13 @@ module BSB
     end
 
     private
+
     def data_hash
-      @data_hash ||= JSON.parse(IO.read(File.expand_path("../../config/bsb_db.json", __FILE__)))
+      @data_hash ||= JSON.parse(IO.read(File.expand_path('../config/bsb_db.json', __dir__)))
     end
 
     def bank_list
-      @bank_list ||= JSON.parse(IO.read(File.expand_path("../../config/bsb_bank_list.json", __FILE__)))
+      @bank_list ||= JSON.parse(IO.read(File.expand_path('../config/bsb_bank_list.json', __dir__)))
     end
   end
 end
