+++ lib/bsb/version.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module BSB
-  VERSION = '0.0.16'.freeze
+  VERSION = '0.0.16'
 end
