+++ lib/bsb_number_validator.rb
@@ -1,10 +1,10 @@
+# frozen_string_literal: true
+
 require 'bsb'
 require 'active_model'
 
 class BsbNumberValidator < ActiveModel::EachValidator
   def validate_each(record, attribute, value)
-    unless BSB.lookup(value)
-      record.errors.add(attribute, :invalid)
-    end
+    record.errors.add(attribute, :invalid) unless BSB.lookup(value)
   end
 end
